% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/helpers.R
\name{find_label}
\alias{find_label}
\title{Find variable label}
\usage{
find_label(.data, .var)
}
\arguments{
\item{.data}{A data.frame containig the labels}

\item{.var}{A character string of length one with the name of the variable}
}
\value{
The variable lable of the requested variable.
}
\description{
Find variable label
}

