% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/plot_factors.R
\name{plot_factors}
\alias{plot_factors}
\title{Autmated bar plots}
\usage{
plot_factors(.data, .labels = NULL)
}
\arguments{
\item{.data}{A \code{data.frame}.}

\item{.labels}{A second \code{data.frame}, containing the variables with
corresponding labels to set titles}
}
\value{
Plots as side effect.
}
\description{
This functions returns bar plots for all the factors in a data set.
}

