% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/data.R
\docType{data}
\name{refactored}
\alias{refactored}
\title{Dataset containing the answers from the survey}
\format{An object of class \code{tbl_df} (inherits from \code{tbl}, \code{data.frame}) with 36 rows and 32 columns.}
\usage{
refactored
}
\description{
Dataset containing the answers from the survey
}
\keyword{datasets}

